Ask gaussian blur kernel size to user
User can insert the gaussian kernel size and who many passes can be perform
Using wrap around strategy for first and last row/column of the image
Pass CodeMaid to remove unnecessary lines and spaces

+++ BMPReader.h
@@ -38,18 +38,18 @@ class BMPReader
         BMPHeader      mBMPHeader;
 
     public:
-        BMPReader           (const std::string &);
-        ~BMPReader          (void);
+        BMPReader(const std::string &);
+        ~BMPReader(void);
 
-        void    setSize         (int, int);
-        void    setAngle        (float);
-        void    addBrightness   (int);
+        void    setSize(int, int);
+        void    setAngle(float);
+        void    addBrightness(int);
 
-        void    showHeader      (void);
-        void    toNegative	    (void);
-        void    toGray		    (void);
+        void    showHeader(void);
+        void    toNegative(void);
+        void    toGray(void);
 
-        void    saveFile	    (const std::string &);
-        void    blurFilter      (int, float);
-        void    sobelFilter     (float);
+        void    saveFile(const std::string &);
+        void    blurFilter(int, int);
+        void    sobelFilter(float);
 };
